msm: kgsl: Add ADRENO_DEPRECATED to identify unsupported targets
Very soon we will start removing support for some pre-production and
obsolete targets from the driver. In order to prevent confusion
add a new flag to identify GPU targets as deprecated so that we don't
accidently match as a different target or otherwise behave badly without
a clue to the poor developer trying to figure out whats going on.

Change-Id: Ic0dedbad8fbfbf6b40bdbec24169ad4649edd28f
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>